+++ lib/Pages/More/camera_page.dart
@@ -6,6 +6,29 @@ import 'package:memories_photos/Popups/toast.dart';
 import 'package:memories_photos/main.dart';
 import 'package:memories_photos/settings.dart';
 
+late CameraController controller;
+void initCamera(int cam, Function(int camI) oninit) {
+  controller = CameraController(cameras[cam], ResolutionPreset.max);
+  controller
+    .initialize()
+    .then((_) {
+      oninit(cam);
+    })
+    .catchError((Object e) {
+      if (e is CameraException) {
+        switch (e.code) {
+          case 'CameraAccessDenied':
+            // Handle access errors here.
+            break;
+          default:
+            // Handle other errors here.
+            break;
+        }
+      }
+    });
+}
+
+
 class CameraPage extends StatefulWidget {
   const CameraPage({super.key});
 
@@ -14,126 +37,145 @@ class CameraPage extends StatefulWidget {
 }
 
 class _CameraPageState extends State<CameraPage> {
-  late CameraController controller;
+  static _CameraPageState? instance;
   double minZoom = 1.0;
   double maxZoom = 2.0;
   double zoom = 1.0;
-  var padding = EdgeInsets.all(0);
 
-  void initCameraZoom() async {
+  void onCamUpdates(int camI) async {
     minZoom = await controller.getMinZoomLevel();
     maxZoom = await controller.getMaxZoomLevel();
     setState(() {});
   }
 
-  void initCamera(int cam) {
-    controller = CameraController(cameras[cam], ResolutionPreset.max);
-    controller
-      .initialize()
-      .then((_) {
-        if (!mounted)
-          return;
-        initCameraZoom();
-        setState(() {});
-      })
-      .catchError((Object e) {
-        if (e is CameraException) {
-          switch (e.code) {
-            case 'CameraAccessDenied':
-              // Handle access errors here.
-              break;
-            default:
-              // Handle other errors here.
-              break;
-          }
-        }
-      });
-  }
-
   @override
   void initState() {
     super.initState();
+    instance = this;
     KeepScreenOn.turnOn();
-    initCamera(0);
+    initCamera(0, (camI) => onCamUpdates(camI));
   }
 
   @override
   void dispose() {
     controller.dispose();
     KeepScreenOn.turnOff();
+    instance = null;
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    padding = MediaQuery.paddingOf(context);
     if (!controller.value.isInitialized) {
       return Container();
     }
     return Scaffold(
-      body: Column(
-        mainAxisAlignment: .spaceBetween,
+      body: Stack(
+        alignment: .center,
         children: [
-          SizedBox(height: padding.top),
-          Flexible(child: CameraPreview(controller)),
-          
           Container(
-            padding: .symmetric(horizontal: 10),
-            height: 35,
-            child: ListView.builder(
-              itemCount: cameras.length,
-              scrollDirection: .horizontal,
-              physics: BouncingScrollPhysics(),
-              itemBuilder: (context, index) => OutlinedButton(
-                onPressed: () => initCamera(index), 
-                child: Text(
-                  "${switch (cameras[index].lensDirection) {
-                    .front => "Front",
-                    .back => "Back",
-                    _ => "???"
-                  }} ${switch (cameras[index].lensType) {
-                    .telephoto => "Default",
-                    .wide => "Wild",
-                    .ultraWide => "Ultrawild",
-                    _ => "???"
-                  }}"
-                )
-              ),
+            clipBehavior: .antiAlias,
+            decoration: BoxDecoration(
+              borderRadius: .circular(25 * Settings.rm)
+            ),
+            child: CameraPreview(controller)
+          ),
+          Padding(
+            padding: MediaQuery.paddingOf(context),
+            child: Column(
+              mainAxisAlignment: .spaceBetween,
+              children: [                
+                Container(),
+            
+                Column(
+                  spacing: 5,
+                  children: [
+                    SliderTheme(
+                      data: SliderThemeData(
+                        
+                      ),
+                      child: Slider(
+                        value: zoom,
+                        min: minZoom,
+                        max: maxZoom,
+                        onChanged: (value) {
+                          controller.setZoomLevel(value);
+                          setState(() => zoom = value);
+                        },
+                      ),
+                    ),
+                    _ShutterRow(controller: controller),
+                    _CameraModeRow(),
+                  ],
+                ),
+                
+              ],
             ),
           ),
-
-          Slider(
-            value: zoom,
-            min: minZoom,
-            max: maxZoom,
-            onChanged: (value) async {
-              await controller.setZoomLevel(value);
-              setState(() => zoom = value);
-            },
-          ),
-
-          ShutterRow(controller: controller),
-          SizedBox(height: padding.bottom)
         ],
       ),
     );
   }
 }
 
-class ShutterRow extends StatelessWidget {
-  const ShutterRow({
-    super.key,
-    required this.controller,
-  });
+class _CameraModeRow extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: .symmetric(horizontal: 10),
+      height: 35,
+      child: Row(
+        children: [
+          OutlinedButton(
+            child: Text("Normal"),
+            onPressed: () => initCamera(0, (camI) => _CameraPageState.instance?.onCamUpdates(camI)),
+          ),
+          OutlinedButton(
+            child: Text("wide"),
+            onPressed: () => initCamera(0, (camI) => _CameraPageState.instance?.onCamUpdates(camI)),
+          )
+        ],
+      )
+    );
+  }
+}
+
+class _ShutterRow extends StatelessWidget {
+  const _ShutterRow({required this.controller});
   final CameraController controller;
 
   @override
   Widget build(BuildContext context) {
     return Row(
-      mainAxisAlignment: .spaceAround,
+      mainAxisAlignment: .spaceEvenly,
       children: [
-        IconButton.filled(
-          icon: Icon(Icons.camera, size: 60),
+        IconButton(
+          icon: Icon(Icons.camera_outlined, size: 60),
+          tooltip: "Capture",
+          onPressed: () async {
+            var photo = await controller.takePicture();
+            final t = DateTime.now();
+            var time = "Photo ${t.year}-${t.month}-${t.day} ${t.hour}-${t.minute}-${t.second}.jpg";
+            await photo.saveTo(Settings.appPath + Platform.pathSeparator + time);
+            await File(photo.path).delete();
+            showStyledToast("Saved: $time", context);
+          },
+        ),
+        IconButton(
+          icon: Icon(Icons.camera_outlined, size: 60),
+          tooltip: "Capture",
+          onPressed: () async {
+            var photo = await controller.takePicture();
+            final t = DateTime.now();
+            var time = "Photo ${t.year}-${t.month}-${t.day} ${t.hour}-${t.minute}-${t.second}.jpg";
+            await photo.saveTo(Settings.appPath + Platform.pathSeparator + time);
+            await File(photo.path).delete();
+            showStyledToast("Saved: $time", context);
+          },
+        ),
+        IconButton(
+          icon: Icon(Icons.camera_outlined, size: 60),
+          tooltip: "Capture",
           onPressed: () async {
             var photo = await controller.takePicture();
             final t = DateTime.now();